fix runner

+++ pkg/runner/runner.go
@@ -3,15 +3,18 @@ package runner
 import (
 	"context"
 	"fmt"
-	"github.com/kubeshop/testkube/contrib/executor/jmeter/pkg/parser"
+	"github.com/joshdk/go-junit"
 	"github.com/kubeshop/testkube/pkg/executor/env"
 	"os"
+	"os/user"
 	"path/filepath"
+	"strings"
 
 	"github.com/kubeshop/testkube/pkg/executor/scraper"
 
 	"github.com/pkg/errors"
 
+	outputPkg "github.com/kubeshop/testkube/pkg/executor/output"
 	"github.com/kubeshop/testkube/pkg/executor/scraper/factory"
 
 	"github.com/kubeshop/testkube/pkg/api/v1/testkube"
@@ -47,34 +50,33 @@ type KarateRunner struct {
 	dependency string
 }
 
-func (r *KarateRunner) Run(ctx context.Context, execution testkube.Execution) (result testkube.ExecutionResult, err error) {
-	if r.Scraper != nil {
-		defer r.Scraper.Close() //invoke after return
+func (karateRunner *KarateRunner) Run(ctx context.Context, execution testkube.Execution) (result testkube.ExecutionResult, err error) {
+	if karateRunner.Scraper != nil {
+		defer karateRunner.Scraper.Close() //invoke after return
 	}
 	output.PrintLogf("%s Preparing for test run", ui.IconTruck)
 
-	err = r.Validate(execution)
+	err = karateRunner.Validate(execution)
 	if err != nil {
 		return result, err
 	}
 
 	// check that the datadir exists
-	_, err = os.Stat(r.Params.DataDir)
+	_, err = os.Stat(karateRunner.Params.DataDir)
 	if errors.Is(err, os.ErrNotExist) {
-		output.PrintLogf("%s Datadir %s does not exist", ui.IconCross, r.Params.DataDir)
+		output.PrintLogf("%s Datadir %s does not exist", ui.IconCross, karateRunner.Params.DataDir)
 		return result, errors.Errorf("datadir not exist: %v", err)
 	}
 
-	runPath := filepath.Join(r.Params.DataDir, "repo", execution.Content.Repository.Path)
+	runPath := filepath.Join(karateRunner.Params.DataDir, "repo", execution.Content.Repository.Path)
+
 	projectPath := runPath
-	if execution.Content.Repository != nil && execution.Content.Repository.WorkingDir != "" {
-		runPath = filepath.Join(r.Params.DataDir, "repo", execution.Content.Repository.WorkingDir)
-	}
+
+	// check that pom.xml file exists
 	fileInfo, err := os.Stat(projectPath)
 	if err != nil {
 		return result, err
 	}
-
 	if !fileInfo.IsDir() {
 		output.PrintLogf("%s Using file...", ui.IconTruck)
 
@@ -84,8 +86,16 @@ func (r *KarateRunner) Run(ctx context.Context, execution testkube.Execution) (r
 	}
 	output.PrintLogf("%s Test content checked", ui.IconCheckMark)
 
+	pomXmlPath := filepath.Join(projectPath, "pom.xml")
+
+	_, pomXmlErr := os.Stat(pomXmlPath)
+	if errors.Is(pomXmlErr, os.ErrNotExist) {
+		outputPkg.PrintLog(fmt.Sprintf("%s No pom.xml found", ui.IconCross))
+		return *result.Err(fmt.Errorf("no pom.xml found")), nil
+	}
+
 	//TODO -  Dependency installation if needed?
-	/*out, err := r.installDependencies(runPath)
+	/*out, err := karateRunner.installDependencies(runPath)
 	if err != nil {
 		return result, errors.Errorf("cypress binary install error: %v\n\n%s", err, out)
 	}*/
@@ -100,43 +110,121 @@ func (r *KarateRunner) Run(ctx context.Context, execution testkube.Execution) (r
 		}
 		envVars = append(envVars, fmt.Sprintf("%s=%s", value.Name, value.Value))
 	}
+	// pass additional executor arguments/flags to maven
+	args := []string{}
+	args = append(args, execution.Args...)
+
+	if execution.VariablesFile != "" {
+		outputPkg.PrintLog(fmt.Sprintf("%s Creating settings.xml file", ui.IconWorld))
+		settingsXML, err := createSettingsXML(projectPath, execution.VariablesFile)
+		if err != nil {
+			outputPkg.PrintLog(fmt.Sprintf("%s Could not create settings.xml", ui.IconCross))
+			return *result.Err(fmt.Errorf("could not create settings.xml")), nil
+		}
+		outputPkg.PrintLog(fmt.Sprintf("%s Successfully created settings.xml", ui.IconCheckMark))
+		args = append(args, "--settings", settingsXML)
+	}
+
+	// workaround for https://github.com/eclipse/che/issues/13926
+	os.Unsetenv("MAVEN_CONFIG")
+	currentUser, err := user.Current()
+	if err == nil && currentUser.Name == "maven" {
+		args = append(args, "-Duser.home=/home/maven")
+	}
+
+	if execution.Content.Repository != nil && execution.Content.Repository.WorkingDir != "" {
+		runPath = filepath.Join(karateRunner.Params.DataDir, "repo", execution.Content.Repository.WorkingDir)
+	}
+
+	mavenCommand := "mvn"
+	goal := "test"
+	if !strings.EqualFold(goal, "project") {
+		// use the test subtype as goal or phase when != project
+		// in case of project there is need to pass additional args
+		args = append(args, goal)
+	}
+
 	//run tests while ignoring execution error in case of failed tests
+	for i := 0; i < len(args); i++ {
+		output.PrintLogf("%s: DebugMode args_%n = "+args[i], ui.IconMicroscope, i) //todo delete this line
+	}
 
 	output.PrintLogf("%s: DebugMode projectPath = "+projectPath, ui.IconMicroscope) //todo delete this line
 	output.PrintLogf("%s: DebugMode runPath = "+runPath, ui.IconMicroscope)         //todo delete this line
-	out, err := executor.Run(runPath, "mvn test", envManager)
+	out, err := executor.Run(runPath, mavenCommand, envManager, args...)
 	out = envManager.ObfuscateSecrets(out)
 
-	result = MapResultsToExecutionResults(out)
-
-	if r.Params.ScrapperEnabled {
-		if err = scrapeArtifacts(ctx, r, execution); err != nil {
+	if karateRunner.Params.ScrapperEnabled {
+		if err = scrapeArtifacts(ctx, karateRunner, execution); err != nil {
 			return result, err
 		}
+	}
+
+	result = MapResultsToExecutionResults(out, err)
+
+	result.Output = string(out)
+	result.OutputType = "text/plain"
+
+	junitReportPath := filepath.Join(projectPath, "target", "surefire-reports")
+	err = filepath.Walk(junitReportPath, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+		if !info.IsDir() && filepath.Ext(path) == ".xml" {
+			suites, _ := junit.IngestFile(path)
+			for _, suite := range suites {
+				for _, test := range suite.Tests {
+					result.Steps = append(
+						result.Steps,
+						testkube.ExecutionStepResult{
+							Name:     fmt.Sprintf("%s - %s", suite.Name, test.Name),
+							Duration: test.Duration.String(),
+							Status:   MapStatus(test.Status),
+						})
+				}
+			}
+		}
+
+		return nil
+	})
+
+	if err != nil {
+		return *result.Err(err), nil
 	}
 
 	return result, err
 }
 
 func scrapeArtifacts(ctx context.Context, r *KarateRunner, execution testkube.Execution) (err error) {
+	reportDirName := "surefire-reports"
+	compressedDirName := reportDirName + "-zip"
 
-	reportPath := filepath.Join(r.Params.DataDir, "repo", execution.Content.Repository.Path, "target")
+	projectPath := filepath.Join(r.Params.DataDir, "repo", execution.Content.Repository.Path)
 
-	originalName := "target"
-	compressedName := originalName + "-zip"
+	surefireReportPath := filepath.Join(projectPath, "target", reportDirName)
+	compressedDirPath := filepath.Join(projectPath, compressedDirName)
 
-	if _, err := executor.Run(reportPath, "mkdir", nil, compressedName); err != nil {
-		output.PrintLogf("%s Artifact scraping failed: making dir %s", ui.IconCross, compressedName)
+	if _, err := executor.Run(projectPath, "mkdir", nil, compressedDirPath); err != nil {
+		output.PrintLogf("%s Artifact scraping failed: making dir %s", ui.IconCross, compressedDirPath)
 		return errors.Errorf("mkdir error: %v", err)
 	}
 
-	if _, err := executor.Run(reportPath, "zip", nil, compressedName+"/"+originalName+".zip", "-r", originalName); err != nil {
-		output.PrintLogf("%s Artifact scraping failed: zipping reports %s", ui.IconCross, originalName)
+	if _, err := executor.Run(projectPath, "ls", nil); err != nil {
+		output.PrintLogf("%s command ls failed", ui.IconCross)
+		return errors.Errorf("ls error: %v", err)
+	}
+
+	output.PrintLogf("Zipping reportPath: %v", surefireReportPath)
+
+	zipCommand := "zip"
+	//zipArgument := "-r \"" + compressedDirPath + "/" + reportDirName + ".zip\"  \"" + surefireReportPath + "\"" //quoting both paths for arguments
+	if _, err := executor.Run(projectPath, zipCommand, nil, compressedDirPath+"/"+reportDirName+".zip", "-r", surefireReportPath); err != nil {
+		output.PrintLogf("%s Artifact scraping failed: zipping reports %s", ui.IconCross, surefireReportPath)
 		return errors.Errorf("zip error: %v", err)
 	}
 
 	directories := []string{
-		filepath.Join(reportPath, compressedName),
+		compressedDirPath,
 	}
 	output.PrintLogf("Scraping directories: %v", directories)
 
@@ -188,38 +276,41 @@ func (r *KarateRunner) installDependencies(runPath string) (out []byte, err erro
 	return nil, errors.Errorf("Other dependencies than Maven are not supported yet")
 }
 
-func MapResultsToExecutionResults(out []byte) (result testkube.ExecutionResult) {
-	result.Status = testkube.ExecutionStatusPassed
-	//TODO - copies from jmeter mapping, write your own
-	/*if results.HasError {
+func MapResultsToExecutionResults(out []byte, err error) (result testkube.ExecutionResult) {
+	if err == nil {
+		result.Status = testkube.ExecutionStatusPassed
+		outputPkg.PrintLog(fmt.Sprintf("%s Test run successful", ui.IconCheckMark))
+	} else {
 		result.Status = testkube.ExecutionStatusFailed
-		result.ErrorMessage = results.LastErrorMessage
+		result.ErrorMessage = err.Error()
+		outputPkg.PrintLog(fmt.Sprintf("%s Test run failed: %s", ui.IconCross, err.Error()))
+		if strings.Contains(result.ErrorMessage, "exit status 1") {
+			// probably some tests have failed
+			result.ErrorMessage = "build failed with an exception"
+		} else {
+			// maven was unable to run at all
+			return result
+		}
 	}
-
-	result.Output = string(out)
-	result.OutputType = "text/plain"
-
-	for _, r := range results.Results {
-		result.Steps = append(
-			result.Steps,
-			testkube.ExecutionStepResult{
-				Name:     r.Label,
-				Duration: r.Duration.String(),
-				Status:   MapStatus(r),
-				AssertionResults: []testkube.AssertionResult{{
-					Name:   r.Label,
-					Status: MapStatus(r),
-				}},
-			})
-	}*/
-
 	return result
 }
 
-func MapStatus(result parser.Result) string {
-	if result.Success {
+func MapStatus(in junit.Status) string {
+	switch string(in) {
+	case "passed":
 		return string(testkube.PASSED_ExecutionStatus)
+	default:
+		return string(testkube.FAILED_ExecutionStatus)
+	}
+}
+
+// createSettingsXML saves the settings.xml to maven config folder and adds it to the list of arguments
+func createSettingsXML(directory string, content string) (string, error) {
+	settingsXML := filepath.Join(directory, "settings.xml")
+	err := os.WriteFile(settingsXML, []byte(content), 0644)
+	if err != nil {
+		return "", fmt.Errorf("could not create settings.xml: %w", err)
 	}
 
-	return string(testkube.FAILED_ExecutionStatus)
+	return settingsXML, nil
 }
